Export function to create ruby text object
